Explicitly specify if <ul> is 'compact' or 'spaced'
Use a class to explicitly specify if <ul> is 'compact' or 'spaced', rather
than always making it compact by using 0 margin, and then trying to give it
a spaced layout by using <br> or <p> in each <li>.

Previously the class 'spaced' was used for <ul> in few places, but had no
effect, as it was not defined.

In few places <ol> is being used, but the CSS specifies the list-style-type
for all <li> in #main as 'disc' rather than 'decimal', so this was
equivalent to <ul>, so change it to that.