Remove min-width from #footer
Remove min-width and other sizing/positioning elements from #footer.  This
just makes index.html look odd in a narrow window, and I can't see any
reason not to just lay this out naturally.

Future work: it's not very clear if #footer is meant to be used inside or
after #main