Workaround historical cygcheck -p bug
cygcheck -p doesn't correctly application/x-www-form-urlencode strings
containing a literal '+' (a useful metacharacter for grep searches), leaving
them as is, rather than %-encoding them.

Workaround this by not decoding '+' as an encoded ' ' if it looks like this
request is coming from cycheck -p (i.e. text=1).  This should be ok since
' ' would be encoded as '%20' by cygcheck, not that it's very useful...